Refactor NakadiHttpClient into an interface, move all AHC-dependent classes into own sub-package.
