+++ src/main/java/org/zalando/rxnakadi/http/ahc/SubscriptionTarget.java
@@ -1,4 +1,4 @@
-package org.zalando.rxnakadi.http;
+package org.zalando.rxnakadi.http.ahc;
 
 import static java.util.Objects.requireNonNull;
