+++ src/test/java/org/zalando/rxnakadi/rx/dispatch/RxDispatchTest.java
@@ -11,10 +11,9 @@ import static org.hamcrest.Matchers.is;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-
-import static org.zalando.rxnakadi.http.AhcResponseDispatch.contentType;
-import static org.zalando.rxnakadi.http.AhcResponseDispatch.onClientError;
-import static org.zalando.rxnakadi.http.AhcResponseDispatch.statusCode;
+import static org.zalando.rxnakadi.http.ahc.AhcResponseDispatch.contentType;
+import static org.zalando.rxnakadi.http.ahc.AhcResponseDispatch.onClientError;
+import static org.zalando.rxnakadi.http.ahc.AhcResponseDispatch.statusCode;
 import static org.zalando.rxnakadi.rx.dispatch.RxDispatch.dispatch;
 import static org.zalando.rxnakadi.rx.dispatch.RxDispatch.on;
